requirements.txt
selenium_recaptcha_solver==1.9.0
selenium==4.8.0
fake_useragent==1.5.1
passlib==1.7.4
mailtm==0.0.81
faker==26.0.0
Flask==3.0.3
python-dotenv==1.0.1